Freshened saxpath.jar to account for the fix to namespace-prefixes in name-steps.
Name-steps now (hopefully) deal 100% correctly with namespaces.
Removed some wackiness regarding a floating ns-context that
followed the evaluation of the xpath in the target document,
which was Just Plain Weird.

Also, reworked handling of the no-prefix case, and improved
logic when the context node is a document or a namespace node.
Thinking I still have an issue with PIs.  Will address that
next.

Removed bogus tests that attempted paths such as "*:foo/*:bar".